[github] Don't consider PR body in CI skip logic
GitHub pull request edits were not triggering CI when the PR
body/description contained the `[ ci skip ]`/`[ skip ci ]` pattern. This
was an issue, for example, in Dependabot-generated PRs that include
commit messages from upstream dependencies that do include the pattern.
Furthermore, no other CI service or vendor considers the PR body for the
"ci skip" mechanism. This corrects the behavior so that only commit
messages or PR titles are considered when a PR is edited.