Handle 5xx trigger errors
Currently, if build triggering fails with an internal error,
the service responds with an 4xx error. Some consumers (including one of ours) would build a retry mechanism on top of error codes,
and wouldn't retry a 400 error, this would result in "losing"
that build trigger request.

Please note that theoretically an input param validation can also result
in the same Errors field being populated and thus a 5xx error.
We don't see any validation error though in the recent logs,
so we propose to deal with this later as being able to separate between
these errors would require some additional refactoring.